+++ cmd/config_cmd.go
@@ -19,25 +19,14 @@ package main
  */
 
 import (
-	"bufio"
-	"bytes"
 	"fmt"
-	"io"
 	"os"
-	"text/template"
 
-	"github.com/hexops/gotextdiff"
-	"github.com/hexops/gotextdiff/myers"
-	"github.com/hexops/gotextdiff/span"
-	"github.com/manifoldco/promptui"
-	"github.com/synfinatic/aws-sso-cli/sso"
-	"github.com/synfinatic/aws-sso-cli/utils"
+	"github.com/synfinatic/aws-sso-cli/internal/utils"
 )
 
 const (
 	AWS_CONFIG_FILE = "~/.aws/config"
-	CONFIG_PREFIX   = "# BEGIN_AWS_SSO_CLI"
-	CONFIG_SUFFIX   = "# END_AWS_SSO_CLI"
 	CONFIG_TEMPLATE = `# BEGIN_AWS_SSO_CLI
 {{ range $sso, $struct := . }}{{ range $arn, $profile := $struct }}
 [profile {{ $profile.Profile }}]
@@ -48,22 +37,22 @@ credential_process = {{ $profile.BinaryPath }} -u {{ $profile.Open }} -S "{{ $pr
 `
 )
 
-var VALID_CONIFG_OPEN []string = []string{
+var VALID_CONFIG_OPEN []string = []string{
 	"clip",
 	"exec",
 	"open",
 }
 
 type ConfigCmd struct {
-	Diff  bool   `kong:"help='Print a diff of changes to the config file instead of modifying it'"`
-	Open  string `kong:"help='Specify how to open URLs: [clip|exec|open]'"`
-	Print bool   `kong:"help='Print profile entries instead of modifying config file'"`
+	Diff  bool   `kong:"help='Print a diff of changes to the config file instead of modifying it',xor='action'"`
 	Force bool   `kong:"help='Write a new config file without prompting'"`
+	Open  string `kong:"help='Specify how to open URLs: [clip|exec|open]'"`
+	Print bool   `kong:"help='Print profile entries instead of modifying config file',xor='action'"`
 }
 
 func (cc *ConfigCmd) Run(ctx *RunContext) error {
 	open := ctx.Settings.ConfigUrlAction
-	if utils.StrListContains(ctx.Cli.Config.Open, VALID_CONIFG_OPEN) {
+	if utils.StrListContains(ctx.Cli.Config.Open, VALID_CONFIG_OPEN) {
 		open = ctx.Cli.Config.Open
 	}
 
@@ -80,142 +69,17 @@ func (cc *ConfigCmd) Run(ctx *RunContext) error {
 		return err
 	}
 
-	templ := configTemplate()
+	f, err := utils.NewFileEdit(CONFIG_TEMPLATE, profiles)
+	if err != nil {
+		return err
+	}
 
 	if ctx.Cli.Config.Print {
-		return templ.Execute(os.Stdout, profiles)
-	}
-	return updateConfig(ctx, templ, profiles)
-}
-
-// generateNewFile generates the contents of a new config file
-func generateNewFile(ctx *RunContext, templ *template.Template, profiles *sso.ProfileMap) ([]byte, error) {
-	var output bytes.Buffer
-	w := bufio.NewWriter(&output)
-
-	// read & write up to the prefix
-	configFile := awsConfigFile()
-	input, err := os.Open(configFile) // output/tempFile is now the input!
-	if err != nil {
-		return []byte{}, err
-	}
-	defer input.Close()
-
-	r := bufio.NewReader(input)
-
-	line, err := r.ReadString('\n')
-	for err == nil && line != fmt.Sprintf("%s\n", CONFIG_PREFIX) {
-		if _, err = w.WriteString(line); err != nil {
-			return []byte{}, err
-		}
-		line, err = r.ReadString('\n')
+		return f.Template.Execute(os.Stdout, profiles)
 	}
 
-	endOfFile := false
-	if err != nil && err != io.EOF {
-		return []byte{}, err
-	} else if err == io.EOF {
-		// Reached EOF before finding our CONFIG_PREFIX
-		endOfFile = true
-	}
-
-	// write our template out
-	if err = templ.Execute(w, profiles); err != nil {
-		return []byte{}, err
-	}
-
-	if !endOfFile {
-		line, err = r.ReadString('\n')
-		// consume our entries and the suffix
-		for err == nil && line != fmt.Sprintf("%s\n", CONFIG_SUFFIX) {
-			line, err = r.ReadString('\n')
-		}
-
-		// if not EOF or other error, read & write the config until EOF
-		if err == nil {
-			// read until error
-			line, err = r.ReadString('\n')
-			for err == nil {
-				if _, err = w.WriteString(line); err != nil {
-					return []byte{}, err
-				}
-				line, err = r.ReadString('\n')
-			}
-			if err != io.EOF {
-				return []byte{}, err
-			}
-		}
-	}
-	w.Flush()
-
-	return output.Bytes(), nil
-}
-
-func configTemplate() *template.Template {
-	templ, err := template.New("profile").Parse(CONFIG_TEMPLATE)
-	if err != nil {
-		log.Panicf("Unable to parse config template: %s", err.Error())
-	}
-
-	return templ
-}
-
-// updateConfig calculates the diff
-func updateConfig(ctx *RunContext, templ *template.Template, profiles *sso.ProfileMap) error {
-	// open our config file
-	configFile := awsConfigFile()
-	inputBytes, err := os.ReadFile(configFile)
-	if err != nil {
-		return err
-	}
-
-	outputBytes, err := generateNewFile(ctx, templ, profiles)
-	if err != nil {
-		return err
-	}
-
-	diff, err := diffBytes(inputBytes, outputBytes, configFile, "new_config.yaml")
-	if err != nil {
-		return err
-	}
-
-	if len(diff) == 0 {
-		// do nothing if there is no diff
-		log.Infof("No changes to made to %s", configFile)
-		return nil
-	}
-
-	if ctx.Cli.Config.Diff {
-		fmt.Printf("%s", diff)
-		return nil
-	}
-
-	if !ctx.Cli.Config.Force {
-		fmt.Printf("The following changes are proposed to %s:\n%s\n\n",
-			utils.GetHomePath("~/.aws/config"), diff)
-		label := "Modify config file with proposed changes?"
-		sel := promptui.Select{
-			Label:        label,
-			Items:        []string{"No", "Yes"},
-			HideSelected: false,
-			Stdout:       &bellSkipper{},
-			Templates: &promptui.SelectTemplates{
-				Selected: fmt.Sprintf(`%s: {{ . | faint }}`, label),
-			},
-		}
-
-		_, val, err := sel.Run()
-		if err != nil {
-			return err
-		}
-
-		if val != "Yes" {
-			fmt.Printf("Aborting.")
-			return nil
-		}
-	}
-
-	return os.WriteFile(configFile, outputBytes, 0600)
+	oldConfig := awsConfigFile()
+	return f.UpdateConfig(ctx.Cli.Config.Diff, ctx.Cli.Config.Force, oldConfig)
 }
 
 // awsConfigFile returns the path the the users ~/.aws/config
@@ -227,12 +91,4 @@ func awsConfigFile() string {
 	}
 	log.Debugf("path = %s", path)
 	return path
-}
-
-// diffBytes generates a diff between two byte arrays
-func diffBytes(aBytes, bBytes []byte, aName, bName string) (string, error) {
-	edits := myers.ComputeEdits(span.URIFromPath(aName), string(aBytes), string(bBytes))
-	diff := gotextdiff.ToUnified(aName, bName, string(aBytes), edits)
-	log.Debugf("diff:\n%v", diff)
-	return fmt.Sprintf("%s", diff), nil
 }