+++ cmd/static_cmd.go
@@ -32,9 +32,9 @@ import (
 	//	"github.com/manifoldco/promptui"
 	"github.com/manifoldco/promptui"
 	"github.com/synfinatic/aws-sso-cli/awsconfig"
+	"github.com/synfinatic/aws-sso-cli/internal/utils"
 	"github.com/synfinatic/aws-sso-cli/sso"
 	"github.com/synfinatic/aws-sso-cli/storage"
-	"github.com/synfinatic/aws-sso-cli/utils"
 	"github.com/synfinatic/gotable"
 )