+++ internal/helper/zshrc.sh
@@ -1,0 +1,1 @@
+complete -o nospace -C {{ .Executable }} aws-sso