+++ internal/utils/bellskipper.go
@@ -1,4 +1,4 @@
-package main
+package utils
 
 /*
  * AWS SSO CLI
@@ -23,7 +23,7 @@ import (
 )
 
 /*
- * bellSkipper implements an io.WriteCloser that skips the terminal bell
+ * BellSkipper implements an io.WriteCloser that skips the terminal bell
  * character (ASCII code 7), and writes the rest to os.Stderr. It is used to
  * replace readline.Stdout, that is the package used by promptui to display the
  * prompts.
@@ -31,11 +31,11 @@ import (
  * This is a workaround for the bell issue documented in
  * https://github.com/manifoldco/promptui/issues/49#issuecomment-573814976
  */
-type bellSkipper struct{}
+type BellSkipper struct{}
 
 // Write implements an io.WriterCloser over os.Stderr, but it skips the terminal
 // bell character.
-func (bs *bellSkipper) Write(b []byte) (int, error) {
+func (bs *BellSkipper) Write(b []byte) (int, error) {
 	const charBell = 7 // c.f. readline.CharBell
 	if len(b) == 1 && b[0] == charBell {
 		return 0, nil
@@ -44,6 +44,6 @@ func (bs *bellSkipper) Write(b []byte) (int, error) {
 }
 
 // Close implements an io.WriterCloser over os.Stderr.
-func (bs *bellSkipper) Close() error {
+func (bs *BellSkipper) Close() error {
 	return os.Stderr.Close()
 }