+++ sso/cache.go
@@ -27,7 +27,7 @@ import (
 	"time"
 
 	// "github.com/davecgh/go-spew/spew"
-	"github.com/synfinatic/aws-sso-cli/utils"
+	"github.com/synfinatic/aws-sso-cli/internal/utils"
 )
 
 const CACHE_VERSION = 3